build: update config
