+++ examples/starter-additional/.storybook/preview.js
@@ -1,4 +1,4 @@
-import { setCustomElementsManifest } from '@web/storybook-prebuilt/web-components.js';
+import { setCustomElementsManifest } from '@storybook/web-components';
 import '@muonic/muon/js/scoped-custom-element-registry.min.js';
 import cem from '../dist/custom-elements.json';
