+++ examples/starter-additional/package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "muon-serve",
+    "start": "storybook dev -p 8000",
     "build": "rollup -c rollup.config.mjs"
   },
   "keywords": [],
