+++ examples/starter-tokens/.storybook/preview.js
@@ -1,6 +1,5 @@
-import { setCustomElementsManifest } from '@web/storybook-prebuilt/web-components';
+import { setCustomElementsManifest } from '@storybook/web-components';
 import '@muonic/muon/js/scoped-custom-element-registry.min.js';
-// import '../dist/muon.min.css';
 
 const cem= await (
   await fetch(new URL('../dist/custom-elements.json', import.meta.url))
