+++ examples/starter-validation/.storybook/preview.js
@@ -1,15 +1,13 @@
-import { setCustomElementsManifest } from '@web/storybook-prebuilt/web-components.js';
+import { setCustomElementsManifest } from '@storybook/web-components';
 import '@muonic/muon/js/scoped-custom-element-registry.min.js';
 import cem from '../dist/custom-elements.json';
 
-// import '../dist/muon.min.css';
-
 setCustomElementsManifest(cem);
 
-export const parameters = {
-  axe: {
-    disabledRules: [
-      "html-has-lang"
-    ],
-  },
+export const preview = {
+  parameters: {
+    a11y: {
+      test: 'error',
+    }
+  }
 };
