+++ examples/starter/mnx-components/cta/story.js
@@ -3,7 +3,10 @@ import setup from '@muonic/muon/storybook/stories';
 
 const details = setup('testcta', MnxTestCta, 'mnx');
 
-export default details.defaultValues;
+export default {
+  ...details.defaultValues,
+  title: 'MNX Components/CTA'
+};
 
 export const Standard = (args) => details.template(args, (args) => args.text);
 Standard.args = { text: 'Highpoint', type: 'text' };
