+++ examples/starter/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "muon-serve",
+    "start": "storybook dev -p 8000",
     "build": "rollup -c"
   },
   "keywords": [],
