+++ packages/muon/components/inputter/src/inputter-styles.slotted.css
@@ -1,4 +1,4 @@
-@import "./inputter-extends.css";
+@import "inputter-extends.css";
 @import "../../../css/accessibility.css";
 
 light-dom {
